moved save of GroupElementYear into CommonBaseForm and removed overriding ModelForm.save()
