+++ base/tests/forms/education_group/test_common.py
@@ -26,10 +26,13 @@ from unittest.mock import patch
 from base.forms.education_group.common import EducationGroupModelForm, CommonBaseForm
 from base.forms.education_group.mini_training import MiniTrainingModelForm, MiniTrainingForm
 from base.models.education_group import EducationGroup
+from base.models.education_group_type import EducationGroupType
 from base.models.education_group_year import EducationGroupYear
 
 from base.models.entity_version import EntityVersion
+from base.models.group_element_year import GroupElementYear
 from base.tests.factories.authorized_relationship import AuthorizedRelationshipFactory
+from base.tests.factories.group_element_year import GroupElementYearFactory
 from django.test import TestCase
 from base.models.enums import organization_type, education_group_categories
 from base.tests.factories.academic_year import AcademicYearFactory, create_current_academic_year
@@ -163,6 +166,34 @@ class TestCommonBaseFormSave(TestCase):
         self.assertEqual(EducationGroup.objects.all().count(), 1)
 
         self._assert_all_fields_correctly_saved(created_education_group_year)
+
+    @patch('base.models.education_group_type.find_authorized_types', return_value=EducationGroupType.objects.all())
+    def test_create_with_parent(self, mock_find_authorized_types):
+        parent = EducationGroupYearFactory()
+
+        form = MiniTrainingForm(data=self.post_data, parent=parent)
+        self.assertTrue(form.is_valid(), form.errors)
+        created_education_group_year = form.save()
+
+        group_element_year = GroupElementYear.objects.filter(parent=parent, child_branch=created_education_group_year)
+        self.assertTrue(group_element_year.exists())
+
+    @patch('base.models.education_group_type.find_authorized_types', return_value=EducationGroupType.objects.all())
+    def test_update_with_parent_when_existing_group_element_year(self, mock_find_authorized_types):
+        parent = EducationGroupYearFactory()
+
+        entity_version = MainEntityVersionFactory()
+        initial_educ_group_year = EducationGroupYearFactory(administration_entity=entity_version.entity)
+
+        GroupElementYearFactory(parent=parent, child_branch=initial_educ_group_year)
+        initial_count = GroupElementYear.objects.all().count()
+
+        form = MiniTrainingForm(data=self.post_data, instance=initial_educ_group_year, parent=parent)
+        self.assertTrue(form.is_valid(), form.errors)
+        form.save()
+
+        # Assert existing GroupElementYear is reused.
+        self.assertEqual(initial_count, GroupElementYear.objects.all().count())
 
 
 def _get_valid_post_data(category):
